move test data to tests folder
