+++ src/tests/test-app-specs/auction.arc32.json
